Update and rename example.txt to data/potential-pupils.txt
Contains 3 male and 3 female examples. 1 of each is within the parameters (at the upper and lower bounds) and since the isMale and isFemale functions are near identical, the working male line (line 1) is at the upper bounds of "school age" (18) and the working female line is at the lower bounds (5).  This tests extreme valid data. 

Alongside that, there's non-working male and non-working female line. One of which is too high, the other is too low. This tests invalid data (at both ends of the spectrum, both too high and too low). 

Finally, erroneuos data is tested with the blank line. This should be reflected by the erroneuos.txt file having a blank line inbetween the 4 invalid peices of data.